+++ src/main/java/com/trie/ecommerce/security/CustomerUserDetails.java
@@ -1,0 +1,51 @@
+package com.trie.ecommerce.security;
+
+import com.trie.ecommerce.entity.Customer;
+import com.trie.ecommerce.enums.UserRole;
+import lombok.Getter;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Collection;
+import java.util.List;
+
+@Getter
+public class CustomerUserDetails implements UserDetails {
+
+    private final Long id;
+    private final String name;
+    private final String email;
+    private final String password;
+    private final UserRole role;
+
+    public CustomerUserDetails(Customer customer) {
+        this.id = customer.getId();
+        this.name = customer.getName();
+        this.email = customer.getEmail();
+        this.password = customer.getPassword();
+        this.role = customer.getRole();
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return List.of(new SimpleGrantedAuthority("ROLE_" + role.name()));
+    }
+
+    @Override
+    public String getUsername() {
+        return email;
+    }
+
+    @Override
+    public boolean isAccountNonExpired() { return true; }
+
+    @Override
+    public boolean isAccountNonLocked() { return true; }
+
+    @Override
+    public boolean isCredentialsNonExpired() { return true; }
+
+    @Override
+    public boolean isEnabled() { return true; }
+}